requirements.txt
Flask==2.2.3
Flask-RESTX==0.5.1
Flask-SQLAlchemy==2.5.1
Flask-JWT-Extended==4.4.4
Flask-CORS==3.0.10
python-decouple==3.6
passlib==1.7.4
Werkzeug==2.2.3
